Drop Page.section attribute
This attribute goes back to an older design, where
the 'edit page' workflow was different: the edit() method did
part of the work of defining what text you meant to edit. In
the current workflow, text() - which replaced the deprecated
edit() - is only for retrieving text, and save() is supposed
to do all the work of editing the page.

The 'section' attribute is therefore obsolete and dangerous.