use 'exec' in py3-compatible manner
per https://docs.python.org/2/reference/simple_stmts.html ,
as exec is a function not a statement in py3, the py2 version
has been set to allow the subsequent statement to be a tuple,
so we can invoke it like this to make it both py2 and py3
compatible. Without this, byte-compiling the file fails under
py3.