add 'end' arg for PageList
this is sometimes useful; for instance I have a use case for
getting all pages in a given namespace that are alphabetically
between two specific points (because their names are in a
predictable format that involves alphabetically-sorted dates).
'end' is added as the last kwarg to avoid breaking any current
uses that rely on the kwarg ordering (thanks dan for pointing
that out).