Move most tool config from setup.cfg to pyproject.toml
The latest incarnation of bumpversion is bump-my-version, which
considers .cfg configuration "deprecated" and wants it moved
to TOML, so here we go. We may as well move pytest config at
the same time as pytest has supported it for ages. We can drop
the "universal wheel" thing as that was about supporting Python
2 and 3 in the same wheel, and we don't support Python 2 any
more.

flake8 does not support pyproject.toml and it doesn't look like
you can specify aliases for setup.py there either, so we'll leave
those two in setup.cfg for now. Maybe in future we can get rid
of them.

Signed-off-by: Adam Williamson <awilliam@redhat.com>