Add several things to MANIFEST.in that are in current tarballs
The tarballs from current releases have these files in them, but
building with `build` does not include them. I don't know if the
previous maintainer was using setuptools-scm or something like
that, but this should ensure that appropriate files are included
in our new tarballs. I left out things like .gitattributes which
clearly shouldn't be in a release tarball.

Signed-off-by: Adam Williamson <awilliam@redhat.com>