Add container-based integration testing
This adds an integration test that runs a set of mediawiki
containers, configures them enough that the API works, then runs
a few simple tests of logging in, creating, moving and deleting
pages. We also enable this in CI. We can extend this with more
tests later, I think this is a reasonable initial set.

Signed-off-by: Adam Williamson <awilliam@redhat.com>