Drop support for old-style continuation
new-style continuation is available since 1.21, and we are
already forcing it for mw 1.21-1.25 in `api` by adding `continue`
to kwargs if it's not there. Old-style continuation is only
available by passing `rawcontinue=` and is not recommended.
I don't think we need to maintain support for it in the List
class. Theoretically it could maybe still be used if someone is
passing `rawcontinue=""` as an arg to `List` or a subclass of it
(and that overrides the `continue=` that we force in - I'm not
sure which wins), but I can't imagine why anyone would want to do
that.

This improves test coverage because old-style continuation is
uncovered at present. I'd rather remove it than write a test for
it.

See: https://www.mediawiki.org/wiki/API:Continue

Signed-off-by: Adam Williamson <awilliam@redhat.com>