fix: Avoid import conflicts by renaming types.py to _types.py
Renaming `types.py` to `_types.py` to avoid conflicts with Python's
built-in modules. This change addresses some `ImportErrors` such as:

```
ImportError: cannot import name 'MappingProxyType' from partially initialized module 'types' (most likely due to a circular import)
```

As far as I can tell this is not a problem when installing the library
using pip or running tests in tox, only local execution when adding the
source-root to the `PYTHONPATH`. This is e.g. the case in the default
IntelliJ/Pycharm run configuration when testing in scratch files.