Fix warning and crash in Site.revisions (#389)
This should fix the deprecation warning reported in #389, though
I am so far unable to reproduce it. The fix follows the same
pattern as the one for Page.revisions.

This also fixes mwclient throwing a KeyError if you call
Site.revisions with a props argument that does not include
timestamp. This was broken by c23abae , which didn't guard
against 'timestamp' not being in the response.

Signed-off-by: Adam Williamson <awilliam@redhat.com>